Add full NativeImage support for Electron.NET

+++ ElectronNET.API/Entities/BitmapOptions.cs
@@ -6,6 +6,6 @@ namespace ElectronNET.API.Entities
 {
     public class BitmapOptions
     {
-        public double ScaleFactor { get; set; } = 1.0d;
+        public float ScaleFactor { get; set; } = 1.0f;
     }
 }
